requirements.txt
speechrecognition
pyttsx3
python-dotenv
requests
rapidfuzz
pyaudio
flask
openai
pygame
